refactor:miner-issue-repo-section
